+++ src/application.rs
@@ -64,6 +64,7 @@ pub mod drehen;
 #[path = "application/empfänger.rs"]
 pub mod empfänger;
 pub mod farbwahl;
+pub mod flat_map;
 pub mod fonts;
 pub mod geschwindigkeit;
 pub mod icon;